refactor authentication scripts
